+++ packages/fields/src/types/Float/index.js
@@ -1,5 +1,5 @@
 import { Float, MongoFloatInterface, KnexFloatInterface } from './Implementation';
-import { importView } from '@keystone-alpha/build-field-types';
+import { importView } from '@keystone/build-field-types';
 
 export default {
   type: 'Float',