+++ packages/fields/src/types/Integer/index.js
@@ -1,5 +1,5 @@
 import { Integer, MongoIntegerInterface, KnexIntegerInterface } from './Implementation';
-import { importView } from '@keystone-alpha/build-field-types';
+import { importView } from '@keystone/build-field-types';
 
 export default {
   type: 'Integer',