+++ packages/fields/src/types/Location/index.js
@@ -1,5 +1,5 @@
 import { Location, MongoLocationInterface, KnexLocationInterface } from './Implementation';
-import { importView } from '@keystone-alpha/build-field-types';
+import { importView } from '@keystone/build-field-types';
 
 export default {
   type: 'Location',