+++ packages/app-admin-ui/client/components/CreateItemModal.js
@@ -6,7 +6,7 @@ import { useToasts } from 'react-toast-notifications';
 
 import { Button, LoadingButton } from '@arch-ui/button';
 import Drawer from '@arch-ui/drawer';
-import { arrayToObject, captureSuspensePromises, countArrays } from '@keystone/utils';
+import { arrayToObject, captureSuspensePromises, countArrays } from '@keystonejs/utils';
 import { gridSize } from '@arch-ui/theme';
 import { AutocompleteCaptor } from '@arch-ui/input';